+++ pages/data/[[...path]].js
@@ -7,8 +7,15 @@ import {fr} from '@codegouvfr/react-dsfr'
 import Head from '@/components/head'
 import Section from '@/components/section'
 import Page from '@/layouts/main'
-import Data from '@/views/data'
 import sendToTracker, {getDownloadToEventTracker} from '@/lib/util/analytics-tracker'
+import Data, {
+  config as dataConfig,
+  getAlias,
+  getFormatedDate,
+  asyncSendS3,
+  listObjectsRecursively,
+  autorizedPathS3,
+} from '@/views/data'
 
 import ErrorPage from '../_error'
 
@@ -35,195 +42,10 @@ const clientS3 = new S3({
   endpoint: S3_CONFIG_ENDPOINT,
 })
 
-const getFormatedDate = () => {
-  const date = new Date()
-  const dateFormatOptions = {
-    year: 'numeric',
-    month: '2-digit',
-    day: '2-digit',
-    hour: '2-digit',
-    minute: '2-digit',
-    second: '2-digit',
-  }
-  return new Intl.DateTimeFormat('fr', dateFormatOptions)
-    .format(date)
-    .replace(/,/g, '\'')
-    .replace(/ /g, ' ')
-}
-
-const autorizedPathS3 = path => {
-  const isPathIsVisibleFile = !path.startsWith('.')
-
-  return {
-    path,
-    auth: isPathIsVisibleFile
-  }
-}
-
-const asyncSendS3 = (req, res, {params, fileName, metadata}) => new Promise(
-  (resolve, reject) => {
-    res.attachment(params.key)
-    res.setHeader('Content-Length', metadata.ContentLength)
-    res.setHeader('Content-Disposition', `attachment; filename="${fileName}"`)
-
-    /* eslint-disable promise/prefer-await-to-then */
-    clientS3.getObject(params)
-      .then(({Body}) => {
-        Body
-          .on('error', err => {
-            const formattedDate = getFormatedDate()
-            console.warn(`[${formattedDate} - ERROR]`, 'File access error:', err)
-            reject(err)
-          })
-          .on('end', () => {
-            resolve()
-          })
-          .pipe(res)
-      })
-      .catch(err => console.error(err))
-  /* eslint-enable promise/prefer-await-to-then */
-  }
-)
-
-// Reccursivly GET S3 Objects
-async function listObjectsRecursively(prefix, continuationToken) {
-  const params = {
-    Bucket: bucketName,
-    Prefix: prefix || undefined,
-    Delimiter: '/',
-    ContinuationToken: continuationToken,
-  }
-
-  try {
-    const data = await clientS3.listObjectsV2(params)
-
-    if (data.KeyCount <= 0) {
-      return undefined
-    }
-
-    const filesList = (data.Contents || []).map(({Key, LastModified, ETag, Size}) => ({
-      name: Key.replace(new RegExp(`^${prefix}`), ''),
-      path: Key,
-      fileInfo: {
-        date: LastModified.toString(),
-        hash: ETag,
-        size: Size,
-      },
-      isDirectory: false,
-    }))
-    const dirsList = (data.CommonPrefixes || []).map(({Prefix}) => ({
-      name: Prefix.replace(new RegExp(`^${prefix}`), '').replace(/\/$/, ''),
-      path: Prefix,
-      isDirectory: true,
-    }))
-    const nextList = (data.IsTruncated) ? await listObjectsRecursively(prefix, data.NextContinuationToken) : []
-    return [...filesList, ...dirsList, ...nextList]
-  } catch (err) {
-    console.error('Erreur lors de la récupération de la liste des objets :', err)
-  }
-}
-
-const config = {
-  alias: [
-    {
-      parent: 'adresses-cadastre',
-      name: 'latest',
-      target: {
-        action: 'getLatestFromRegExp',
-        params: ['^\\d{4}$']
-      },
-    },
-    {
-      parent: 'adresses-ftth',
-      name: 'latest',
-      target: {
-        action: 'getLatestFromRegExp',
-        params: ['^\\d{4}-T\\d$']
-      },
-    },
-    {
-      parent: 'ban',
-      name: 'adresses-odbl',
-      target: 'adresses',
-      comment: 'Ce repertoire est obsolète et redirige vers le répertoire adresses',
-    },
-    {
-      parent: 'ban/adresses',
-      name: 'weekly',
-      target: {
-        action: 'getLatestFromRegExp',
-        params: ['^\\d{4}-\\d{2}-\\d{2}$']
-      },
-    },
-    {
-      parent: 'ban/export-api-gestion', // OBSOLETE ?
-      name: 'latest',
-      target: {
-        action: 'getLatestFromRegExp',
-        params: ['^\\d{4}-\\d{2}-\\d{2}$']
-      },
-    },
-    {
-      parent: 'contours-administratifs',
-      name: 'latest',
-      target: {
-        action: 'getLatestFromRegExp',
-        params: ['^\\d{4}$']
-      },
-    },
-    {
-      parent: 'fantoir',
-      name: 'latest',
-      target: {
-        action: 'getLatestFromRegExp',
-        params: ['^fantoir-\\d{4}-\\d{2}\\.gz$']
-      },
-    },
-  ]
-}
-
-const aliasAction = {
-  getLatestFromRegExp: (s3Objects, regExpString) => s3Objects
-    .filter(({name}) => ((new RegExp(regExpString)).test(name)))
-    .sort((a, b) => new Date(b.name) - new Date(a.name))[0]
-}
-
-const getAlias = async (aliasesRaw, currentPath) => {
-  const aliases = aliasesRaw && (
-    [...config.alias]
-      .sort((a, b) => `${b.parent}${b.name}`.localeCompare(`${a.parent}${a.name}`))
-      .reverse()
-  )
-  const alias = aliases?.find(({parent}) => (new RegExp(`^${parent}(/|$)`)).test(currentPath)) || null
-
-  if (typeof alias.target === 'string') {
-    return alias
-  }
-
-  if (typeof alias.target === 'object' && alias.target.action) {
-    const s3ObjectPath = [...rootDir, ...alias.parent.replace(/\/$/, '').split('/')].join('/')
-    const s3DirPath = `${s3ObjectPath}/`
-    const s3Objects = await listObjectsRecursively(s3DirPath)
-    const s3data = s3Objects ? [
-      ...(s3Objects || [])
-        .filter(({name}) => autorizedPathS3(name).auth),
-    ] : []
-
-    const targetAlias = aliasAction[alias.target.action](s3data, ...(alias.target?.params || []))
-
-    return {
-      ...alias,
-      target: targetAlias.name,
-    }
-  }
-
-  throw new Error('Alias target should be a string, or an object with an action key')
-}
-
 export async function getServerSideProps(context) {
   const {params, res, req} = context
   const {path: paramPathRaw = []} = params
-  const alias = await getAlias(config?.alias, paramPathRaw.join('/') || '')
+  const alias = await getAlias(clientS3, bucketName)(rootDir, dataConfig?.alias, paramPathRaw.join('/') || '')
 
   const paramPath = alias && (new RegExp(`^${alias.parent}/${alias.name}`)).test(paramPathRaw.join('/'))
     ? paramPathRaw.join('/').replace(new RegExp(`^${alias.parent}/${alias.name}`), `${alias.parent}/${alias.target}`).split('/')
@@ -245,7 +67,7 @@ export async function getServerSideProps(context) {
         downloadFileName: dirPath,
         nbDownload: 1
       }))
-      await asyncSendS3(req, res, {
+      await asyncSendS3(clientS3)(req, res, {
         params: {
           Bucket: bucketName,
           Key: s3ObjectPath,
@@ -265,7 +87,7 @@ export async function getServerSideProps(context) {
   } catch {
     // ERROR > PATH SHOULD BE DIRECTORY
     const s3DirPath = `${s3ObjectPath}/`
-    const s3Objects = await listObjectsRecursively(s3DirPath)
+    const s3Objects = await listObjectsRecursively(clientS3, bucketName)(s3DirPath)
     if (s3Objects) {
       const s3data = [
         ...(s3Objects || [])
