+++ views/data/index.js
@@ -1,1 +1,10 @@
 export {default} from './data'
+export {
+  getAlias,
+  getFormatedDate,
+  asyncSendS3,
+  listObjectsRecursively,
+  autorizedPathS3,
+} from './helper'
+
+export {default as config} from './data-config.json'
